Use an explicit (int) cast for _gold_to_donation. That should fix Eino's compile time warning. :)
git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7782 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/religion.cc
@@ -5975,7 +5975,7 @@ static piety_gain_t _sacrifice_one_item_noncount( const item_def& item)
 
 static int _gold_to_donation(int gold)
 {
-    return ((gold * log((double)gold)) / MAX_PIETY);
+    return (int) ((gold * log((double)gold)) / MAX_PIETY);
 }
 
 void offer_items()
